separate logic and presentation from event links block
